+++ exception/TelegramBot.java
@@ -6,7 +6,7 @@ import java.net.URL;
 /**
  * TelegramBot
  * @author Yazmin
- * @version 08/05/2020
+ * @version 17/07/2020
  */
 
 public class TelegramBot {
@@ -19,21 +19,32 @@ public class TelegramBot {
         this.group = groupId;
     }
 
+    /**
+     * Method to adda message
+     * @param message to send
+     */
     public void addMessage(String message) {
         this.message += message + "%0A";
     }
 
+    /**
+     * Method to send to telegram
+     */
     public void send() {
         try {
             String completeURL = this.url + this.group + "&text=" + this.message;
             new URL(completeURL).openConnection().getInputStream();
-        } catch (IOException e) {
-            e.printStackTrace();
+        } catch (IOException ioException) {
+            new Exception().log(ioException);
         } finally {
             this.message = "";
         }
     }
 
+    /**
+     * Method to send a exception to telegram
+     * @param exception message to send
+     */
     public static void sendToTelegram(String exception) {
         TelegramBot telegramBot = new TelegramBot("ErroresPP");
         telegramBot.addMessage(exception);
