+++ gui/administrator/fxml/FXMLUpdateCoordinator.fxml
@@ -10,6 +10,7 @@
 <?import javafx.scene.control.RadioButton?>
 <?import java.net.URL?>
 <?import javafx.scene.layout.Pane?>
+<?import javafx.scene.control.Tooltip?>
 
 <AnchorPane id="AnchorPane" prefHeight="406.0" prefWidth="400.0" styleClass="mainFxmlClass"
             xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1"
@@ -23,8 +24,8 @@
             <Font name="Arial Bold" size="24.0"/>
          </font>
       </Label>
-      <Button fx:id="btnLogOut" layoutX="494.0" layoutY="13.0" mnemonicParsing="false" onAction="#logOutAdministrator"
-              style="-fx-background-color: WHITE;" text="Cerrar Sesión">
+      <Button fx:id="btnLogOut" layoutX="494.0" layoutY="13.0" mnemonicParsing="false"
+              onAction="#logOutAdministrator" style="-fx-background-color: WHITE;" text="Cerrar Sesión">
          <font>
             <Font name="Arial Bold" size="12.0"/>
          </font>
@@ -73,7 +74,11 @@
    <TextField fx:id="tfName" layoutX="266.0" layoutY="131.0" prefHeight="25.0" prefWidth="208.0"/>
    <TextField fx:id="tfLastName" layoutX="266.0" layoutY="161.0" prefHeight="25.0" prefWidth="208.0"/>
    <TextField fx:id="tfEmail" layoutX="266.0" layoutY="240.0" prefHeight="25.0" prefWidth="208.0"/>
-   <TextField fx:id="tfAlternateEmail" layoutX="266.0" layoutY="271.0" prefHeight="25.0" prefWidth="208.0"/>
+   <TextField fx:id="tfAlternateEmail" layoutX="266.0" layoutY="271.0" prefHeight="25.0" prefWidth="208.0">
+      <tooltip>
+         <Tooltip text="Debe ser difernente al anterior"/>
+      </tooltip>
+   </TextField>
    <TextField fx:id="tfPhone" layoutX="266.0" layoutY="301.0" prefHeight="25.0" prefWidth="208.0"/>
    <Button fx:id="btnUpdate" layoutX="295.0" layoutY="360.0" mnemonicParsing="false" onAction="#updateCoordinator"
            style="-fx-background-color: #59ad55;" text="Modificar" textFill="WHITE">
