+++ gui/administrator/fxml/FXMLUpdateTeacher.fxml
@@ -10,6 +10,7 @@
 <?import javafx.scene.control.RadioButton?>
 <?import java.net.URL?>
 <?import javafx.scene.layout.Pane?>
+<?import javafx.scene.control.Tooltip?>
 
 <AnchorPane id="AnchorPane" prefHeight="406.0" prefWidth="603.0" styleClass="mainFxmlClass"
             xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1"
@@ -23,8 +24,8 @@
             <Font name="Arial Bold" size="24.0"/>
          </font>
       </Label>
-      <Button fx:id="btnLogOut" layoutX="494.0" layoutY="13.0" mnemonicParsing="false" onAction="#logOutAdministrator"
-              style="-fx-background-color: WHITE;" text="Cerrar Sesión">
+      <Button fx:id="btnLogOut" layoutX="494.0" layoutY="13.0" mnemonicParsing="false"
+              onAction="#logOutAdministrator" style="-fx-background-color: WHITE;" text="Cerrar Sesión">
          <font>
             <Font name="Arial Bold" size="12.0"/>
          </font>
@@ -73,7 +74,11 @@
    <TextField fx:id="tfName" layoutX="271.0" layoutY="132.0" prefHeight="25.0" prefWidth="208.0"/>
    <TextField fx:id="tfLastName" layoutX="271.0" layoutY="162.0" prefHeight="25.0" prefWidth="208.0"/>
    <TextField fx:id="tfEmail" layoutX="271.0" layoutY="241.0" prefHeight="25.0" prefWidth="208.0"/>
-   <TextField fx:id="tfAlternateEmail" layoutX="271.0" layoutY="272.0" prefHeight="25.0" prefWidth="208.0"/>
+   <TextField fx:id="tfAlternateEmail" layoutX="271.0" layoutY="272.0" prefHeight="25.0" prefWidth="208.0">
+      <tooltip>
+         <Tooltip text="Debe ser diferente al anterior"/>
+      </tooltip>
+   </TextField>
    <TextField fx:id="tfPhone" layoutX="271.0" layoutY="302.0" prefHeight="25.0" prefWidth="208.0"/>
    <Button fx:id="btnUpdate" layoutX="188.0" layoutY="355.0" mnemonicParsing="false" onAction="#updateTeacher"
            style="-fx-background-color: #59ad55;" text="Modificar" textFill="WHITE">
