+++ gui/coordinator/fxml/FXMLRegisterPractitioner.fxml
@@ -10,6 +10,7 @@
 <?import javafx.scene.control.RadioButton?>
 <?import java.net.URL?>
 <?import javafx.scene.layout.Pane?>
+<?import javafx.scene.control.Tooltip?>
 
 <AnchorPane id="AnchorPane" prefHeight="452.0" prefWidth="603.0" styleClass="mainFxmlClass"
             xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1"
@@ -67,7 +68,11 @@
    <TextField fx:id="tfName" layoutX="217.0" layoutY="83.0" prefHeight="25.0" prefWidth="208.0"/>
    <TextField fx:id="tfLastName" layoutX="216.0" layoutY="112.0" prefHeight="25.0" prefWidth="208.0"/>
    <TextField fx:id="tfEmail" layoutX="219.0" layoutY="233.0" prefHeight="25.0" prefWidth="208.0"/>
-   <TextField fx:id="tfAlternateEmail" layoutX="219.0" layoutY="262.0" prefHeight="25.0" prefWidth="208.0"/>
+   <TextField fx:id="tfAlternateEmail" layoutX="219.0" layoutY="262.0" prefHeight="25.0" prefWidth="208.0">
+      <tooltip>
+         <Tooltip text="Debe ser diferente al correo"/>
+      </tooltip>
+   </TextField>
    <TextField fx:id="tfPhone" layoutX="219.0" layoutY="292.0" prefHeight="25.0" prefWidth="208.0"/>
    <Button fx:id="btnRegister" layoutX="196.0" layoutY="412.0" mnemonicParsing="false" onAction="#registerPractitioner"
            style="-fx-background-color: #59ad55;" text="Aceptar" textFill="WHITE">
@@ -94,8 +99,16 @@
             <Font name="Arial" size="10.0"/>
          </font>
       </Label>
-      <TextField fx:id="tfEnrollment" layoutX="138.0" layoutY="8.0" prefHeight="25.0" prefWidth="208.0"/>
-      <TextField fx:id="tfPassword" layoutX="139.0" layoutY="41.0" prefHeight="25.0" prefWidth="208.0"/>
+      <TextField fx:id="tfEnrollment" layoutX="138.0" layoutY="8.0" prefHeight="25.0" prefWidth="208.0">
+         <tooltip>
+            <Tooltip text="La S debe ser mayúscula"/>
+         </tooltip>
+      </TextField>
+      <TextField fx:id="tfPassword" layoutX="139.0" layoutY="41.0" prefHeight="25.0" prefWidth="208.0">
+         <tooltip>
+            <Tooltip text="Debe tener al menos 10 caracteres"/>
+         </tooltip>
+      </TextField>
       <Label layoutX="14.0" layoutY="13.0" text="Matrícula">
          <font>
             <Font name="Arial" size="14.0"/>
@@ -107,7 +120,11 @@
          <Font name="Arial" size="14.0"/>
       </font>
    </Label>
-   <TextField fx:id="tfCredits" layoutX="219.0" layoutY="173.0" prefHeight="25.0" prefWidth="208.0"/>
+   <TextField fx:id="tfCredits" layoutX="219.0" layoutY="173.0" prefHeight="25.0" prefWidth="208.0">
+      <tooltip>
+         <Tooltip text="Debe tener al menos el 70% de créditos"/>
+      </tooltip>
+   </TextField>
    <Label fx:id="lbTerm" layoutX="220.0" layoutY="146.0">
       <font>
          <Font name="Arial" size="14.0"/>
