+++ logic/ValidateUser.java
@@ -11,6 +11,11 @@ import java.util.regex.Pattern;
 
 public class ValidateUser extends ValidateGeneral {
 
+    /**
+     *Method to validate name
+     * @param name from User
+     * @return whether it is valid or not
+     */
     public boolean validateNameUser (String name) {
         Pattern pattern = Pattern
                 .compile("[A-Za-zñÑáéíóúÁÉÍÓÚ]{2,30}");
@@ -19,6 +24,11 @@ public class ValidateUser extends ValidateGeneral {
         return result;
     }
 
+    /**
+     * Method to create a correct proper name
+     * @param name from User
+     * @return Correct proper name
+     */
     public String createCorrectProperName (String name) {
         String[] splitName = name.split("");
         for (int index = 0; index < splitName.length; index++) {
@@ -32,30 +42,50 @@ public class ValidateUser extends ValidateGeneral {
         return correctName;
     }
 
-    public boolean validateUserName (String word) {
+    /**
+     *Method to validate username
+     * @param userName from User
+     * @return whether it is valid or not
+     */
+    public boolean validateUserName (String userName) {
         Pattern pattern = Pattern
                 .compile("[A-Za-zñÑáéíóúÁÉÍÓÚ0-9]{10,50}");
-        Matcher mather = pattern.matcher(word);
+        Matcher mather = pattern.matcher(userName);
         boolean result = mather.find();
         return result;
     }
 
-    public boolean validatePassword (String word) {
+    /**
+     * Method to validate a password
+     * @param password from User
+     * @return whether it is valid or not
+     */
+    public boolean validatePassword (String password) {
         Pattern pattern = Pattern
                 .compile("[A-Za-zñÑáéíóúÁÉÍÓÚ0-9]{10,20}");
-        Matcher mather = pattern.matcher(word);
+        Matcher mather = pattern.matcher(password);
         boolean result = mather.find();
         return result;
     }
 
-    public boolean validateEmpty (String works) {
+    /**
+     * Method to validate that a field is not empty
+     * @param word to validate
+     * @return whether it is valid or not
+     */
+    public boolean validateEmpty (String word) {
         boolean result=false;
-        if(works.length() > 0){
+        if(word.length() > 0){
             result=true;
         }
         return result;
     }
 
+    /**
+     * Method to validate a enrollment
+     * @param enrollment from practitioner
+     * @return whether it is valid or not
+     */
     public boolean validateEnrollment (String enrollment) {
         boolean isValidEnrollment;
         Pattern pattern = Pattern
@@ -65,6 +95,11 @@ public class ValidateUser extends ValidateGeneral {
         return isValidEnrollment;
     }
 
+    /**
+     * Method to know if a practitioner has the necessary credits
+     * @param credits from practitioner
+     * @return whether it is valid or not
+     */
     public boolean validateCreditsPractitioner (String credits) {
         boolean validate = true;
         if(Integer.parseInt(credits) < 250 || Integer.parseInt(credits) > 347) {
